requirements.txt
Django
djangorestframework
requests
djangorestframework-simplejwt